sexywrap: don't reuse iscsi_context:s
When clean_up() a target we kept the iscsi_context in the hope that it can be
reused for the next connection.  Unfortunately there's bug in libiscsi 1.13
which disables data transmission after logout and this flag is not cleared.
So let's forget the old iscsi_context and create a new one in open() for the
time being.